pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddiscribe"
version = "1.0.0"
description = "Local AI-powered Reddit analysis and writing assistant for non-English users"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
